requirements.txt
discord.py
dotenv
python-dotenv
psycopg2-binary